# Parallelize the slowest part of the rolling indexer