@ndushay 2023-10-17
notifying @justinlittman and @andrewjbtw that the only negative impact we have thought of is that by parallelizing the rolling reindexer, there will be fewer cores on -a for on demand indexing (and new object indexing) ... but -b is unaffected.  So keep an eye out for unacceptable indexing times for stuff.