@andrewjbtw 2023-10-17
Presumably now that we've left Fedora, the possible negative impact of taking the core part of the system down with too much activity is less of a risk.