@mjgiarlo 2023-10-18
@jmartin-sul 💬 
> based on a quick trace through the code, i'm wondering if there's an additional easy win to be gained in terms of speed based on the speed improvement possibilities mentioned in the PR description, assuming my read of the code is correct.

Yes, good call! Made that change. It's now an ~8x improvement. 😄 
 
> even with that concern, my read is that this change doesn't hurt things in terms of correctness. and given the huge speed improvement and the expectation that even parallelization alone could provide a big speed boost, i'm going to assume that the speed improvement is real and that the test result isn't a random fluke due to the sample size of 1.

Between yesterday and today, @ndushay and I tested this over ~15 times in stage and prod, so the sample size is slightly bigger and the results were pretty consistent, for what it's worth!
 
🍻