@mjgiarlo 2023-10-31
@ndushay still 2: https://github.com/sul-dlss/dor_indexing_app/blob/main/config/settings.yml#L4